Fix naming for screen code usage
In the functions that plot actual character to the screen the naming
suggested it would handle PETSCII codes, where in fact is was dealing
with screen codes. These changes make that more clear.
This could be a breaking change for anyone who has been using
fc_plotPetsciiChar in the past.